refactor: extract reviewer assignment engine
Move pure reviewer-assignment logic into a dedicated
`reviewer_assignment_engine` module and keep
`reviewer_assignment.py` focused on snapshot/model integration.

This preserves current assignment behavior while creating the seam
for round-based rescoring and offline policy simulation.

prepared with codex
