+++ docs/design-decisions/037-reviewer-assignment-policy-simulation-and-priority-planning.md
@@ -318,6 +318,11 @@
   - Follow-up plan refinement:
     - prefer splitting pure engine logic into a dedicated `reviewer_assignment_engine.py`
     - keep trace requirements semantic rather than over-constraining exact shape during early refactors
+- 2026-03-15:
+  - Chunk 1 started:
+    - extracted pure assignment primitives and batch execution into `qb_site/analyzer/services/reviewer_assignment_engine.py`
+    - rewired `reviewer_assignment.py` to act as the snapshot/model integration layer over the engine
+    - preserved current batch semantics, including one-shot ranking, so iterative rescoring remains future work in chunk 2
 
 ## References
 - `docs/design-decisions/README.md`
