+++ qb_site/analyzer/services/__init__.py
@@ -6,12 +6,14 @@ from .dependencies import DependencyRebuildResult, body_hash, parse_dependency_n
 from .reviewer_assignment import (
     AssignmentStatistics,
     AreaStatsBuilder,
+    PRAssignmentPriority,
     ReviewerAssignmentBuilder,
     ReviewerProfile,
     ReviewerSuggestionResult,
     build_reviewer_catalog,
     collect_assignment_statistics,
     compute_area_stats,
+    rank_prs_for_assignment,
     suggest_reviewer_for_pr,
     suggest_reviewers_many,
 )
@@ -44,10 +46,12 @@ __all__ = [
     "ReviewerAssignmentBuilder",
     "ReviewerProfile",
     "AssignmentStatistics",
+    "PRAssignmentPriority",
     "ReviewerSuggestionResult",
     "build_reviewer_catalog",
     "collect_assignment_statistics",
     "compute_area_stats",
+    "rank_prs_for_assignment",
     "suggest_reviewer_for_pr",
     "suggest_reviewers_many",
     "ReviewerAttentionItem",
